TextureAtlas is working.  Full changes in description
- Added texture atlas
- Added com.namone.blocks package
- Added some block 'objects' (dirt, grass, stone)

The texture atlas should be fully working. I tested it and was able to
draw multiple, different, textures. I also had each texture ID, when
called, create a new instance of it's object (ie. Dirt()). There are no
properties for the blocks yet; but I figured I would add it.